Removed old redundant middleware and restructured AspNetCore namespace. Fixes #1323
