+++ core/Piranha.AspNetCore/Extensions/PiranhaHtmlExtensions.cs
@@ -9,15 +9,28 @@
  */
 
 using System.Text;
+using System.Text.RegularExpressions;
 using Microsoft.AspNetCore.Html;
 using Piranha.AspNetCore.Services;
+using Piranha.Extend;
 using Piranha.Models;
 
 /// <summary>
 /// Extension class with html helper methods.
 /// </summary>
-public static class HtmlExtensions
+public static class PiranhaHtmlExtensions
 {
+    /// <summary>
+    /// Converts the type name of the block into a pretty
+    /// css class name.
+    /// </summary>
+    /// <param name="block">The current block</param>
+    /// <returns>The css class name</returns>
+    public static string CssName(this Block block)
+    {
+        return ClassNameToWebName(block.GetType().Name);
+    }
+
     /// <summary>
     /// Generates meta tags for the given content.
     /// </summary>
@@ -85,13 +98,25 @@ public static class HtmlExtensions
         return !string.IsNullOrWhiteSpace(content.MetaTitle) ? content.MetaTitle : content.Title;
     }
 
-    public static string OgTitle(IMeta content)
+    private static string OgTitle(IMeta content)
     {
         return !string.IsNullOrWhiteSpace(content.OgTitle) ? content.OgTitle : MetaTitle(content);
     }
 
-    public static string OgDescription(IMeta content)
+    private static string OgDescription(IMeta content)
     {
         return !string.IsNullOrWhiteSpace(content.OgDescription) ? content.OgDescription : content.MetaDescription;
+    }
+
+    /// <summary>
+    /// Converts a standard camel case class name to a lowercase
+    /// string with each word separated with a dash, suitable
+    /// for use in views.
+    /// </summary>
+    /// <param name="str">The camel case string</param>
+    /// <returns>The converted string</returns>
+    private static string ClassNameToWebName(string str)
+    {
+        return Regex.Replace(str, "([A-Z])", " $1", RegexOptions.Compiled).Trim().Replace(" ", "-").ToLower();
     }
 }
