+++ core/Piranha.AspNetCore/Helpers/IMediaHelper.cs
@@ -12,8 +12,11 @@ using Piranha.Extend.Fields;
 using Piranha.Extend.Blocks;
 using Piranha.Models;
 
-namespace Piranha.AspNetCore.Services
+namespace Piranha.AspNetCore.Helpers
 {
+    /// <summary>
+    /// Helper for manipulating media files.
+    /// </summary>
     public interface IMediaHelper
     {
         /// <summary>
