+++ core/Piranha.AspNetCore/Helpers/RequestHelper.cs
@@ -8,32 +8,32 @@
  *
  */
 
-namespace Piranha.AspNetCore.Services
+namespace Piranha.AspNetCore.Helpers
 {
     /// <summary>
     /// The request helper provides information regarding the
     /// current request.
     /// </summary>
-    public interface IRequestHelper
+    public sealed class RequestHelper : IRequestHelper
     {
         /// <summary>
         /// Gets/sets the current hostname.
         /// </summary>
-        string Host { get; set; }
+        public string Host { get; set; }
 
         /// <summary>
         /// Gets/sets the current port.
         /// </summary>
-        int? Port { get; set; }
+        public int? Port { get; set; }
 
         /// <summary>
         /// Gets/sets the current scheme.
         /// </summary>
-        string Scheme { get; set; }
+        public string Scheme { get; set; }
 
         /// <summary>
         /// Gets/sets the requested raw url.
         /// </summary>
-        string Url { get; set; }
+        public string Url { get; set; }
     }
 }
