+++ core/Piranha.AspNetCore/MiddlewareBase.cs
@@ -10,7 +10,6 @@
 
 using Microsoft.AspNetCore.Http;
 using Microsoft.Extensions.Logging;
-using System;
 using System.Threading.Tasks;
 using Piranha.AspNetCore.Services;
