+++ core/Piranha.AspNetCore/Services/IModelLoader.cs
@@ -15,10 +15,16 @@ using Piranha.Models;
 
 namespace Piranha.AspNetCore.Services
 {
+    /// <summary>
+    /// The model loader is used for retrieving content models with
+    /// built in permission checks for the current user.
+    /// </summary>
     public interface IModelLoader
     {
         /// <summary>
-        /// Gets the specified page model for the given user.
+        /// Gets the specified page model for the given user. If the
+        /// user doesn't have access to the requested page an
+        /// UnauthorizedAccessException is thrown.
         /// </summary>
         /// <param name="id">The unique id</param>
         /// <param name="user">The current user</param>
@@ -28,7 +34,9 @@ namespace Piranha.AspNetCore.Services
         Task<T> GetPageAsync<T>(Guid id, ClaimsPrincipal user, bool draft = false) where T : PageBase;
 
         /// <summary>
-        /// Gets the specified post model for the given user.
+        /// Gets the specified post model for the given user. If the
+        /// user doesn't have access to the requested post an
+        /// UnauthorizedAccessException is thrown.
         /// </summary>
         /// <param name="id">The unique id</param>
         /// <param name="user">The current user</param>
