+++ examples/MvcWeb/Startup.cs
@@ -13,10 +13,12 @@ using Microsoft.AspNetCore.Hosting;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
+using Microsoft.OpenApi.Models;
 using Piranha;
 using Piranha.Data.EF.SQLite;
 using Piranha.AspNetCore.Identity.SQLite;
-using Microsoft.OpenApi.Models;
+using Piranha.AttributeBuilder;
+using Piranha.Local;
 
 namespace MvcWeb
 {
@@ -26,28 +28,26 @@ namespace MvcWeb
         // For more information on how to configure your application, visit https://go.microsoft.com/fwlink/?LinkID=398940
         public void ConfigureServices(IServiceCollection services)
         {
-            services.AddLocalization(options =>
-                options.ResourcesPath = "Resources"
-            );
-            services.AddControllersWithViews();
-            services.AddRazorPages()
-                .AddPiranhaManagerOptions();
+            services.AddPiranha(options =>
+            {
+                options.AddRazorRuntimeCompilation = true;
 
-            services.AddPiranha();
-            services.AddPiranhaApplication();
-            services.AddPiranhaFileStorage();
-            services.AddPiranhaImageSharp();
-            services.AddPiranhaManager();
-            services.AddPiranhaTinyMCE();
-            services.AddPiranhaApi();
+                options.UseFileStorage(naming: FileStorageNaming.UniqueFolderNames);
+                options.UseImageSharp();
+                options.UseManager();
+                options.UseTinyMCE();
+                options.UseMemoryCache();
 
-            services.AddPiranhaEF<SQLiteDb>(options =>
-                options.UseSqlite("Filename=./piranha.mvcweb.db"));
-            services.AddPiranhaIdentityWithSeed<IdentitySQLiteDb>(options =>
-                options.UseSqlite("Filename=./piranha.mvcweb.db"));
+                options.UseEF<SQLiteDb>(db =>
+                    db.UseSqlite("Filename=./piranha.mvcweb.db"));
+                options.UseIdentityWithSeed<IdentitySQLiteDb>(db =>
+                    db.UseSqlite("Filename=./piranha.mvcweb.db"));
 
-            services.AddMemoryCache();
-            services.AddPiranhaMemoryCache();
+                options.UseSecurity(o =>
+                {
+                    o.UsePermission("Subscriber");
+                });
+            });
 
             services.AddSwaggerGen(options =>
             {
@@ -78,20 +78,13 @@ namespace MvcWeb
             App.CacheLevel = Piranha.Cache.CacheLevel.Full;
 
             // Build content types
-            new Piranha.AttributeBuilder.PageTypeBuilder(api)
-                .AddType(typeof(Models.BlogArchive))
-                .AddType(typeof(Models.StandardPage))
-                .AddType(typeof(Models.TeaserPage))
+            new ContentTypeBuilder(api)
+                .AddAssembly(typeof(Startup).Assembly)
                 .Build()
                 .DeleteOrphans();
-            new Piranha.AttributeBuilder.PostTypeBuilder(api)
-                .AddType(typeof(Models.BlogPost))
-                .Build()
-                .DeleteOrphans();
-            new Piranha.AttributeBuilder.SiteTypeBuilder(api)
-                .AddType(typeof(Models.StandardSite))
-                .Build()
-                .DeleteOrphans();
+
+            // Configure editor
+            Piranha.Manager.Editor.EditorConfig.FromFile("editorconfig.json");
 
             /**
              *
@@ -102,22 +95,11 @@ namespace MvcWeb
             System.Globalization.CultureInfo.DefaultThreadCurrentUICulture = cultureInfo;
              */
 
-            // Register middleware
-            app.UseStaticFiles();
-            app.UsePiranha();
-            app.UseRouting();
-            app.UseAuthentication();
-            app.UseAuthorization();
-            app.UsePiranhaIdentity();
-            app.UsePiranhaManager();
-            app.UsePiranhaTinyMCE();
-            app.UseEndpoints(endpoints =>
+            app.UsePiranha(options =>
             {
-                endpoints.MapControllerRoute(
-                    name: "default",
-                    pattern: "{controller=Home}/{action=Index}/{id?}");
-
-                endpoints.MapPiranhaManager();
+                options.UseManager();
+                options.UseTinyMCE();
+                options.UseIdentity();
             });
 
             Seed.RunAsync(api).GetAwaiter().GetResult();
